django/requirements.txt
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
pandas==1.3.1
boto3==1.18.10
django-storages==1.11.1